reduced number of components
